Move the SuspensionQ and friends into its own source unit
To reduce size/bloat in scheduler.rs
